+++ sdk/batch/batch/src/index.ts
@@ -122,8 +122,6 @@ export type {
   OutputFileUploadConfig,
   OutputFileUploadCondition,
   BatchTaskConstraints,
-  AuthenticationTokenSettings,
-  BatchAccessScope,
   BatchJobPreparationTask,
   BatchJobReleaseTask,
   BatchPoolInfo,